#24 refactor: 디렉토리 구조 변경, Wrapper 스타일 변경
