+++ src/components/pages/Solution/index.tsx
