+++ src/components/pages/StudyInfo/SessionGrid/index.tsx
