+++ src/components/pages/StudyInfo/StudyInfo.tsx
