+++ src/components/pages/StudyInfo/StudyInfoHeader/StudySettingMenu/index.tsx
