+++ src/components/pages/StudyInfo/StudyInfoHeader/index.tsx
@@ -3,7 +3,7 @@ import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 import GitHubIcon from "@mui/icons-material/GitHub";
 
-import Header from "../../../components/Header";
+import Header from "../../../blocks/Header";
 import Profile from "../../../blocks/Profile";
 import Wrapper from "../../../blocks/Wrapper";
 import StudySettingMenu from "./StudySettingMenu";
