+++ src/components/pages/StudyInfo/StudyInfoMiddleBar/SearchTextField/index.tsx
