+++ src/components/pages/StudyInfo/StudyInfoMiddleBar/SessionAdder/index.tsx
