+++ src/components/pages/StudyInfo/StudyInfoMiddleBar/SortSplitButton/index.tsx
