+++ src/components/pages/StudyInfo/StudyInfoMiddleBar/index.tsx
