+++ src/components/pages/StudyInfo/index.tsx
