+++ src/hoc/Layout/index.tsx
@@ -3,7 +3,7 @@ import { Container, Stack, CssBaseline } from "@mui/material";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
 
 import { AuthProvider } from "../../hoc/AuthContext";
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/blocks/Navbar";
 
 const baseTheme = createTheme({
   palette: {
